Laravel Project
open the project without public.home in the url using htaccess.
Make the Master Layout and then extends it in the blade templates.
Make the Logout functionality.
Apply the Front page layout as par the client given theme using the Gradient color.
Display the Login section when Click on the Login/Register button.
Make the Footer well.
And Chnages the small designs issues...